+++ biblioteca-back/src/main/java/co/edu/uniandes/g5/bibliotecas/entities/RecursoEntity.java
@@ -16,9 +16,9 @@ import uk.co.jemos.podam.common.PodamExclude;
 @Entity
 public abstract class RecursoEntity extends BaseEntity implements Serializable{
     
-    public static final int VIDEO = 1;
-    public static final int LIBRO = 2;
-    public static final int SALA = 3;
+    public static final Long VIDEO = 1L;
+    public static final Long LIBRO = 2L;
+    public static final Long SALA = 3L;
     
     @PodamExclude
     @OneToMany(mappedBy = "recurso")
@@ -37,13 +37,13 @@ public abstract class RecursoEntity extends BaseEntity implements Serializable{
     @ManyToOne
     private co.edu.uniandes.g5.bibliotecas.entities.BibliotecaEntity biblioteca;
    
-    private int tipoRecurso;
+    private Long tipoRecurso;
 
-    public int getTipoRecurso() {
+    public Long getTipoRecurso() {
         return tipoRecurso;
     }
 
-    public void setTipoRecurso(int tipoRecurso) {
+    public void setTipoRecurso(Long tipoRecurso) {
         this.tipoRecurso = tipoRecurso;
     }
